Separated the tests to 3 sets
